unified local_runtests and runtests with test_suites.jl ; added iris and adult tests
