+++ test/classification/iris.jl
@@ -3,10 +3,7 @@
 
 @testset "iris.jl" begin
 
-import DelimitedFiles
-
-download("https://archive.ics.uci.edu/ml/machine-learning-databases/iris/iris.data", "iris.csv")
-iris = DelimitedFiles.readdlm("iris.csv", ',')
+iris = DelimitedFiles.readdlm("data/iris.csv", ',')
 
 features = iris[:, 1:4]
 labels = iris[:, 5]
